+++ src/lambdaisland/glogi.clj
@@ -2,7 +2,7 @@
 
 (defn- log-expr [form level keyvals]
   (let [keyvals-map (apply array-map keyvals)
-        formatter (::formatter keyvals-map 'pr-str)]
+        formatter (::formatter keyvals-map 'identity)]
     `(log ~(::logger keyvals-map (str *ns*))
           ~level
           (~formatter