+++ CHANGELOG.md
@@ -2,9 +2,15 @@
 
 ## Added
 
-## Fixed
+- `lambdaisland.glogi.console` provides an alternative to `goog.debug.Console`,
+  with the main benefit that it will log full data structures, thus playing
+  nicely with cljs-devtools
 
 ## Changed
+
+- The default formatter is now `identity` instead of `pr-str`. This way we
+  preserve full data structures until the last minute. Note that goog.debug.Logger
+  will stringify the message, unless care is taken for it not to.
 
 # 0.0-13 (2019-06-10 / 0668ebe)